+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Add `CHANGELOG.md` ([#27](https://github.com/salcode/wp-fast-login/issues/27))
+- Change from `wp_localize_script()` to `wp_add_inline_script()` for compatibility with WordPress 6.5 ([#26](https://github.com/salcode/wp-fast-login/issues/26))
 
 ## 0.1.0
 - Initial release